# 3rike-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/loans": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Apply for a loan (requires credit score \u003e= 500)",
                "parameters": [
                    {
                        "description": "Loan",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Loan"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Loan"
                        }
                    }
                }
            }
        },
        "/api/loans/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Get loan by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Loan"
                        }
                    }
                }
            }
        },
        "/api/loans/{id}/repay": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Make a loan repayment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Loan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Repayment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "amount_usdc": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Loan"
                        }
                    }
                }
            }
        },
        "/api/payments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Record a driver weekly repayment",
                "parameters": [
                    {
                        "description": "Payment",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Payment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Payment"
                        }
                    }
                }
            }
        },
        "/api/payments/driver/{driverID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Get all payments for a driver",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver ID",
                        "name": "driverID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Payment"
                            }
                        }
                    }
                }
            }
        },
        "/api/savings/deposit": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "savings"
                ],
                "summary": "Deposit USDC into driver savings account",
                "parameters": [
                    {
                        "description": "Deposit",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "amount_usdc": {
                                    "type": "number"
                                },
                                "driver_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Savings"
                        }
                    }
                }
            }
        },
        "/api/savings/{driverID}/balance": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "savings"
                ],
                "summary": "Get savings balance for a driver",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver ID",
                        "name": "driverID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Savings"
                        }
                    }
                }
            }
        },
...
        "/api/yield/investor/{investorID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "yield"
                ],
                "summary": "Get all yield payouts for an investor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Investor ID",
                        "name": "investorID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.YieldPayout"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Loan": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "principal_usdc": {
                    "type": "number"
                },
                "remaining_usdc": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/domain.LoanStatus"
                },
                "weekly_repayment": {
                    "type": "number"
                }
            }
        },
        "domain.LoanStatus": {
            "type": "string",
            "enum": [
                "active",
                "repaid",
                "defaulted"
            ],
            "x-enum-varnames": [
                "LoanActive",
                "LoanRepaid",
                "LoanDefaulted"
            ]
        },
        "domain.Payment": {
            "type": "object",
            "properties": {
                "amount_local": {
                    "type": "number"
                },
                "amount_usdc": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/domain.PaymentStatus"
                },
                "tricycle_id": {
                    "type": "integer"
                },
                "week_number": {
                    "type": "integer"
                }
            }
        },
        "domain.PaymentStatus": {
            "type": "string",
            "enum": [
                "pending",
                "confirmed",
                "failed"
            ],
            "x-enum-varnames": [
                "PaymentPending",
                "PaymentConfirmed",
                "PaymentFailed"
            ]
        },
...
        "domain.Savings": {
            "type": "object",
            "properties": {
                "balance_usdc": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.YieldPayout": {
            "type": "object",
            "properties": {
                "amount_usdc": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "fraction_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "investor_id": {
                    "type": "integer"
                },
                "week_number": {
                    "type": "integer"
                }
            }
        }
    }
}